Remove "renderChatMessage" hook, move contents to `ChatMessage#getHTML`
There's currently an upstream bug/limitation in which the scroll-to-bottom behavior happens after `ChatMessage#getHTML` is called but before "renderChatMessage" hooks complete
