+++ src/module/chat-message/listeners/index.ts
@@ -1,2 +1,6 @@
+export { ChatCards } from "./cards";
 export { DamageButtons } from "./damage-buttons";
 export { DamageTaken } from "./damage-taken";
+export { DegreeOfSuccessHighlights } from "./degree-of-success";
+export { SetAsInitiative } from "./set-as-initiative";
+export { MessageTooltips } from "./tooltips";
